// frontend/src/generated/GoogleSuggestions.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { GoogleSuggestion } from "./GoogleSuggestion";

export type GoogleSuggestions = { "status": "ok", suggestions: Array<GoogleSuggestion>, } | { "status": "missing_credentials" } | { "status": "timeout" } | { "status": "error" };